Vastly improved collisions.
Hotbox poly's now precalculate surface normals and corner normals. In a collision, the surface normal and corner normals are averaged for a better collision normal.
Overlap is now calculated for all intersections. This is now used as a spring constant for collisions, as well as prioritising which intersection to use in a multiple intersection collision. The above fixes seems to have stamped out ships getting stuck together.
Some rotation shorcuts have been added to Utility, and GameConst now has its own file.